requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.4
itsdangerous==0.24
oauthlib==1.0.3
requests==2.9.1
requests-oauthlib==0.6.1
tweepy==3.5.0
python-telegram-bot==3.3